requirements.txt
pillow==11.1.0
PyMuPDF==1.25.3
python-docx==1.1.2
docx2pdf==0.1.8
pdf2docx==0.5.8
fpdf2==2.8.2
moviepy==2.2.1
pyyaml==6.0.2
markdown==3.7
html2text==2024.2.25
beautifulsoup4==4.13.3
pypdf==5.3.0
pillow-heif==1.3.0
cairosvg==2.7.1
win11toast==0.35
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1